+++ src/main/java/ma/bservices/dao/IAccidentTravailDao.java
@@ -20,7 +20,7 @@ public interface IAccidentTravailDao {
     public void editAccidentTravail(AccidentTravail a);
     public void remouveAccidentTravail(AccidentTravail a);
     public List<AccidentTravail> allAccidentTravail();
-    public AccidentTravail allAccidentTravailById(Integer id);
+    public AccidentTravail allAccidentTravailById(Long id);
     public List<AccidentTravail> allAccidentTravailByChantier(Chantier c);
     public List<AccidentTravail> allAccidentTravailBySalarie(Salarie s);
